Rename standalone/ and modules/ to pages/
Move page-specific init scripts to web_src/js/pages/ and their CSS
to web_src/css/pages/. Remove the standalone/ directory.
