+++ web_src/js/pages/swagger.ts
