--- web_src/js/standalone/external-render-iframe.ts
@@ -1,4 +1,0 @@
-import '../../css/standalone/external-render-iframe.css';
-import {initExternalRenderIframe} from '../modules/external-render-iframe.ts';
-
-initExternalRenderIframe();
